+++ tests/version/test_helper.py
@@ -17,7 +17,36 @@
 
 import unittest
 
-from pontos.version.helper import safe_version
+from pontos.version.helper import (
+    VersionError,
+    is_version_pep440_compliant,
+    safe_version,
+    strip_version,
+)
+
+
+class IsVersionPep440CompliantTestCase(unittest.TestCase):
+    def test_is_compliant(self):
+        self.assertTrue(is_version_pep440_compliant("1.2.3.dev1"))
+        self.assertTrue(is_version_pep440_compliant("1.2.3.dev0"))
+        self.assertTrue(is_version_pep440_compliant("20.4"))
+        self.assertTrue(is_version_pep440_compliant("1.2"))
+        self.assertTrue(is_version_pep440_compliant("1.2.0a0"))
+        self.assertTrue(is_version_pep440_compliant("1.2.0a1"))
+        self.assertTrue(is_version_pep440_compliant("1.2.0b0"))
+        self.assertTrue(is_version_pep440_compliant("1.2.0b1"))
+
+    def test_is_not_compliant(self):
+        self.assertFalse(is_version_pep440_compliant("1.2.3dev1"))
+        self.assertFalse(is_version_pep440_compliant("1.2.3dev"))
+        self.assertFalse(is_version_pep440_compliant("1.2.3dev0"))
+        self.assertFalse(is_version_pep440_compliant("1.2.3alpha"))
+        self.assertFalse(is_version_pep440_compliant("1.2.3alpha0"))
+        self.assertFalse(is_version_pep440_compliant("1.2.3.a0"))
+        self.assertFalse(is_version_pep440_compliant("1.2.3beta"))
+        self.assertFalse(is_version_pep440_compliant("1.2.3beta0"))
+        self.assertFalse(is_version_pep440_compliant("1.2.3.b0"))
+        self.assertFalse(is_version_pep440_compliant("20.04"))
 
 
 class SafeVersionTestCase(unittest.TestCase):
@@ -52,3 +81,23 @@ class SafeVersionTestCase(unittest.TestCase):
         self.assertEqual(safe_version("1.2"), "1.2")
         self.assertEqual(safe_version("1.2.3"), "1.2.3")
         self.assertEqual(safe_version("22.4"), "22.4")
+
+
+class StripVersionTestCase(unittest.TestCase):
+    def test_version_string_without_v(self):
+        self.assertEqual(strip_version("1.2.3"), "1.2.3")
+        self.assertEqual(strip_version("1.2.3dev"), "1.2.3dev")
+
+    def test_version_string_with_v(self):
+        self.assertEqual(strip_version("v1.2.3"), "1.2.3")
+        self.assertEqual(strip_version("v1.2.3dev"), "1.2.3dev")
+
+
+class VersionErrorTestCase(unittest.TestCase):
+    def test_should_print_message(self):
+        err = VersionError("foo bar")
+        self.assertEqual(str(err), "foo bar")
+
+    def test_should_raise(self):
+        with self.assertRaisesRegex(VersionError, "^foo bar$"):
+            raise VersionError("foo bar")
