+++ tests/version/test_version.py
@@ -19,6 +19,7 @@ import contextlib
 import io
 import unittest
 from pathlib import Path
+from unittest.mock import MagicMock
 
 from pontos.version.helper import VersionError
 from pontos.version.version import VersionCommand
@@ -40,3 +41,13 @@ class VersionCommandTestCase(unittest.TestCase):
             VersionCommand(project_file_path=Path("whatever")).run(
                 args=["show"]
             )
+
+    def test_print_current_version(self):
+        print_mock = MagicMock()
+        VersionCommand.get_current_version = MagicMock(return_value="1.2.3")
+        VersionCommand._print = print_mock  # pylint: disable=protected-access
+
+        cmd = VersionCommand()
+        cmd.print_current_version()
+
+        print_mock.assert_called_with("1.2.3")
